+++ src/components/dialog/RemoveUserDialog.vue
@@ -1,9 +1,9 @@
 <template>
   <q-dialog v-model="show">
     <q-card>
-      <q-card-section class="flex row">
-        <span class="text-h6 text-center">
-          {{ $t('RemoveUserDialog.text.title', { name: user.name }) }}
+      <q-card-section class="flex row justify-center">
+        <span class="text-h6">
+          {{ $t('RemoveUserDialog.text.title', { name: user.username }) }}
         </span>
       </q-card-section>
       <q-form @submit="onSubmit">